requirements.txt

tensorflow>=2.5.0
tensorflow-hub>=0.12.0
tensorflow-datasets>=4.4.0
matplotlib>=3.4.2
numpy>=1.19.5
scikit-learn>=0.24.2
seaborn>=0.11.1
Pillow>=8.2.0
